// severskcable-pwa/src/types/image-response.ts
export type successImageResponse = {
    successImage: string;
}

export type successImageActionType = {
    breakId: string;
    successImage: string;
}

export type registerImageActionType = {
    breakId: string;
    registerImage: string;
}

export type repairingImageActionType = {
    breakId: string;
    repairingImage: string;
}

export type repairCompletedImageActionType = {
    breakId: string;
    repairCompletedImage: string;
}